feat: deleting unnecesary scripts
